+++ docs/changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - integration tests for the CLI
 - docker image for the CLI application
 - has been added the `parse` method to the `SpamAnalyzer` class to parse a single mail, use that instead of the mailparser library to don't see unwanted logs
+- pre-commit hooks to check the code before committing
 
 ### Changed
 
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MailAnalyzer` class has been renamed to `SpamAnalyzer
 - `MailAnalysis` class has no longer methods to decide if a mail is spam or not, those are now in the `SpamAnalyzer` class
 - to make the package really modular, now the classification model can be injected in the `SpamAnalyzer` class (if not provided, the default one will be used)
+- config folder is now created in the standard OS path for configuration files using click
 
 ### Fixed